Implement size limit for hybrid LowLat/Bulk traffic
Some protocols such as HTTP and SSH are normally desired to be
low-latency on upstream; however they can also be used for bulk uploads
thus affecting performance of every other protocols.

Create a special "LowLat" filter that also has a simple higher-bound
limit on packet size. That limit is based on a bitmask so it's not quite
granular; set at 1024.