curfew: Flush connections based on MAC addresses
Curfew activation used to flush the whole conntrack table, causing all
TCP connections on the router to reset. We now dig the ip address
associated with the MAC and flush matching entries instead.

This is a workaround to the fact the user-supplied rules always comes
after the "RELATED,ESTABLISHED" one making rules ineffective for
existing connections (and also reduces the number of rules to go through
on established connections).